remove `KtorGithubAPI` and `RateLimit` singletons
