+++ core/src/com/unciv/logic/github/GithubAPI.kt
@@ -1,10 +1,23 @@
 package com.unciv.logic.github
 
+import com.badlogic.gdx.Files
+import com.badlogic.gdx.files.FileHandle
+import com.unciv.UncivGame
 import com.unciv.json.json
+import com.unciv.logic.UncivKtor
+import com.unciv.logic.UncivShowableException
+import com.unciv.logic.github.Github.repoNameToFolderName
 import com.unciv.logic.github.GithubAPI.parseUrl
+import io.ktor.client.plugins.*
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
 import io.ktor.http.*
+import kotlinx.coroutines.delay
+import java.io.FileFilter
+import java.util.zip.ZipException
+import kotlin.time.Clock
+import kotlin.time.ExperimentalTime
+import kotlin.time.Instant
 
 /**
  *  "Namespace" collects all Github API structural knowledge
@@ -21,6 +34,53 @@ import io.ktor.http.*
 @Suppress("PropertyName")  // We're declaring an external API schema
 object GithubAPI {
     // region URL formatters
+    /**
+     * @see <a href="https://ktor.io/docs/client-default-request.html#url">
+     *          Ktor Client > Developing applications > Requests > Default request > Base URL
+     *      </a>
+     */
+    const val baseUrl = "https://api.github.com"
+
+    /**
+     * Add a bearer token here if needed
+     *
+     * @see <a href="https://github.com/yairm210/Unciv/issues/13951#issuecomment-3326406877">#13951 (comment)</a>
+     */
+    const val bearerToken = ""
+
+    private val client = UncivKtor.client.config {
+        defaultRequest {
+            url(baseUrl)
+            header("X-GitHub-Api-Version", "2022-11-28")
+            header(HttpHeaders.Accept, "application/vnd.github+json")
+            userAgent(UncivGame.getUserAgent("Github"))
+            if (bearerToken.isNotBlank()) bearerAuth(bearerToken)
+        }
+    }
+
+    /**
+     * Make a ktor request handling rate limits automatically
+     */
+    suspend fun request(
+        maxRateLimitedRetries: Int = 3,
+        block: HttpRequestBuilder.() -> Unit,
+    ): HttpResponse {
+        val resp = client.request(block)
+        val rateLimited = consumeRateLimit(resp)
+
+        return if (rateLimited) {
+            if (maxRateLimitedRetries <= 0) return resp
+            return request(maxRateLimitedRetries - 1, block)
+        } else resp
+    }
+
+    private suspend fun paginatedRequest(
+        page: Int, amountPerPage: Int, block: HttpRequestBuilder.() -> Unit
+    ) = request {
+        parameter("page", page)
+        parameter("per_page", amountPerPage)
+        block()
+    }
 
     /** Format a download URL for a branch archive */
     // URL format see: https://docs.github.com/en/repositories/working-with-files/using-files/downloading-source-code-archives#source-code-archive-urls
@@ -33,10 +93,53 @@ object GithubAPI {
 
     /** Format a URL to query a repo tree - to calculate actual size */
     // It's hard to see in the doc this not only accepts a commit SHA, but either branch (used here) or tag names too
-    internal suspend fun Repo.fetchReleaseZip() = KtorGithubAPI.request {
+    internal suspend fun Repo.fetchReleaseZip() = request {
         url("/repos/$full_name/git/trees/$default_branch")
         parameter("recursive", "true")
     }
+
+    /**
+     * Wait for rate limit to end if any and returns true if there was any rate limit
+     */
+    @OptIn(ExperimentalTime::class)
+    private suspend fun consumeRateLimit(resp: HttpResponse): Boolean {
+        if (resp.status != HttpStatusCode.Forbidden && resp.status != HttpStatusCode.TooManyRequests) return false
+
+        val remainingRequests = resp.headers["x-ratelimit-remaining"]?.toIntOrNull() ?: 0
+        if (remainingRequests < 1) return false
+
+        val resetEpoch = resp.headers["x-ratelimit-reset"]?.toLongOrNull() ?: 0
+        delay(Instant.fromEpochSeconds(resetEpoch) - Clock.System.now())
+
+        return true
+    }
+
+    suspend fun fetchGithubReposWithTopic(search: String, page: Int, amountPerPage: Int) =
+        paginatedRequest(page, amountPerPage) {
+            url("/search/repositories")
+            parameter("sort", "stars")
+            parameter("q", "$search${if (search.isEmpty()) "" else "+"} topic:unciv-mod fork:true")
+        }
+
+    suspend fun fetchGithubTopics() = request {
+        url("/search/topics")
+        parameter("sort", "name")
+        parameter("order", "asc")
+
+        /**
+         * `repositories:>1` means ignore unused or practically unused topics
+         */
+        parameter("q", "unciv-mod repositories:>1")
+    }
+
+    suspend fun fetchSingleRepo(owner: String, repoName: String) =
+        request { url("/repos/$owner/$repoName") }
+
+    /**
+     * We are not using KtorGithubAPI here because the URL provided is not an API URL
+     */
+    suspend fun fetchPreviewImageOrNull(modUrl: String, branch: String, ext: String) =
+        UncivKtor.getOrNull("$modUrl/$branch/preview.${ext}") { host = "raw.githubusercontent.com" }
 
     //endregion
     //region responses
@@ -95,7 +198,7 @@ object GithubAPI {
 
             /** Query Github API for [owner]'s [repoName] repository metadata */
             suspend fun query(owner: String, repoName: String): Repo? {
-                val resp = KtorGithubAPI.fetchSingleRepo(owner, repoName)
+                val resp = fetchSingleRepo(owner, repoName)
                 return if (!resp.status.isSuccess()) null
                 else json().fromJson(Repo::class.java, resp.bodyAsText())
             }
@@ -236,4 +339,194 @@ object GithubAPI {
         return this
     }
     //endregion
+
+    /**
+     * Download a mod and extract, deleting any pre-existing version.
+     * @param modsFolder Destination handle of mods folder - also controls Android internal/external.
+     * @param updateProgressPercent A function recieving a download progress percentage (0-100) roughly every 100ms. Call will run under a Coroutine context.
+     * @author **Warning**: This took a long time to get just right, so if you're changing this, ***TEST IT THOROUGHLY*** on _both_ Desktop _and_ Phone.
+     * @return FileHandle for the downloaded Mod's folder or null if download failed.
+     */
+    suspend fun Repo.downloadAndExtract(
+        modsFolder: FileHandle,
+        /** Should accept a number 0-100 */
+        updateProgressPercent: ((Int) -> Unit)? = null
+    ): FileHandle? {
+        var modNameFromFileName = name
+
+        val defaultBranch = default_branch
+        val zipUrl: String
+        val tempName: String
+        if (direct_zip_url.isEmpty()) {
+            val gitRepoUrl = html_url
+            // Initiate download - the helper returns null when it fails
+            zipUrl = getUrlForBranchZip(gitRepoUrl, defaultBranch)
+
+            // Get a mod-specific temp file name
+            tempName = "temp-" + gitRepoUrl.hashCode().toString(16)
+        } else {
+            zipUrl = direct_zip_url
+            tempName = "temp-" + toString().hashCode().toString(16)
+        }
+
+        val resp = UncivKtor.client.get(zipUrl)
+        /**
+         * This block is borrowed from the deleted method `processRequestHandlingRedirects()`
+         * See: https://github.com/yairm210/Unciv/blob/1cb3f94d36009719b63f6d8e9d2e49c1bd594a8f/core/src/com/unciv/logic/github/Github.kt#L197-L216
+         */
+        when {
+            resp.status == HttpStatusCode.NotFound -> return null
+            (resp.status == HttpStatusCode.Forbidden) &&
+                resp.headers["CF-RAY"] != null && resp.headers["cf-mitigated"].orEmpty() == "challenge" ->
+                throw UncivShowableException("Blocked by Cloudflare")
+
+            (resp.status.value in 401..403) || resp.status == HttpStatusCode.ProxyAuthenticationRequired ->
+                throw UncivShowableException("Servers requiring authentication are not supported")
+
+            resp.status.value in 300..499 ->
+                throw UncivShowableException("Unexpected response: [${resp.bodyAsText()}]")
+
+            resp.status.value in 500..599 ->
+                throw UncivShowableException("Server failure: [${resp.bodyAsText()}}]")
+
+            !resp.status.isSuccess() ->
+                throw UncivShowableException("Unknown Issue!\nStatus: ${resp.status}\nBody:[${resp.bodyAsText()}}]")
+        }
+
+        val content = resp.bodyAsBytes()
+        if (content.isEmpty()) return null
+
+        val disposition = resp.headers[HttpHeaders.ContentDisposition]
+        modNameFromFileName = parseNameFromDisposition(disposition, modNameFromFileName)
+
+        // Download to temporary zip
+        // If the Content-Length header was missing, fall back to the reported repo size (which should be in kB)
+        // and assume low compression efficiency - bigger mods typically have mostly images.
+        val tempZipFileHandle = modsFolder.child("$tempName.zip")
+        tempZipFileHandle.writeBytes(content, false)
+
+        // prepare temp unpacking folder
+        val unzipDestination = tempZipFileHandle.sibling(tempName) // folder, not file
+        // prevent mixing new content with old - hopefully there will never be cadavers of our tempZip stuff
+        if (unzipDestination.exists())
+            if (unzipDestination.isDirectory) unzipDestination.deleteDirectory() else unzipDestination.delete()
+
+        try {
+            Zip.extractFolder(tempZipFileHandle, unzipDestination)
+        } catch (ex: ZipException) {
+            throw UncivShowableException("That is not a valid ZIP file", ex)
+        }
+
+        val (innerFolder, modName) = resolveZipStructure(unzipDestination, modNameFromFileName)
+
+        // modName can be "$repoName-$defaultBranch"
+        val finalDestinationName = modName.replace("-$defaultBranch", "").repoNameToFolderName()
+        // finalDestinationName is now the mod name as we display it. Folder name needs to be identical.
+        val finalDestination = modsFolder.child(finalDestinationName)
+
+        // prevent mixing new content with old
+        var tempBackup: FileHandle? = null
+        if (finalDestination.exists()) {
+            tempBackup = finalDestination.sibling("$finalDestinationName.updating")
+            finalDestination.renameOrMove(tempBackup)
+        }
+
+        // Move temp unpacked content to their final place
+        finalDestination.mkdirs() // If we don't create this as a directory, it will think this is a file and nothing will work.
+        // The move will reset the last modified time (recursively, at least on Linux)
+        // This sort will guarantee the desktop launcher will not re-pack textures and overwrite the atlas as delivered by the mod
+        for (innerFileOrFolder in innerFolder.list()
+            .sortedBy { file -> file.extension() == "atlas" }) {
+            innerFileOrFolder.renameOrMove(finalDestination)
+        }
+
+        // clean up
+        tempZipFileHandle.delete()
+        unzipDestination.deleteDirectory()
+        if (tempBackup != null)
+            if (tempBackup.isDirectory) tempBackup.deleteDirectory() else tempBackup.delete()
+
+        return finalDestination
+    }
+
+
+    private val parseAttachmentDispositionRegex =
+        Regex("""attachment;\s*filename\s*=\s*(["'])?(.*?)\1(;|$)""")
+
+    private fun parseNameFromDisposition(disposition: String?, default: String): String {
+        fun String.removeZipExtension() = removeSuffix(".zip").replace('.', ' ')
+        if (disposition == null) return default.removeZipExtension()
+        val match = parseAttachmentDispositionRegex.matchAt(disposition, 0)
+            ?: return default.removeZipExtension()
+        return match.groups[2]!!.value.removeZipExtension()
+    }
+
+    private fun FileHandle.renameOrMove(dest: FileHandle) {
+        // Gdx tries a java rename for Absolute and External, but NOT for Local - rectify that
+        if (type() == Files.FileType.Local) {
+            // See #5346: renameTo 'unpacks' a source folder if the destination exists and is an
+            // empty folder, dropping the name of the source entirely.
+            // Safer to ask for a move to the not-yet-existing full resulting path instead.
+            if (isDirectory)
+                if (file().renameTo(dest.child(name()).file())) return
+                else
+                    if (file().renameTo(dest.file())) return
+        }
+        moveTo(dest)
+    }
+
+    private val goodFolders =
+        listOf("Images", "jsons", "maps", "music", "sounds", "Images\\..*", "scenarios", ".github")
+            .map { Regex(it, RegexOption.IGNORE_CASE) }
+    private val goodFiles = listOf(
+        ".*\\.atlas",
+        ".*\\.png",
+        "preview.jpg",
+        ".*\\.md",
+        "Atlases.json",
+        ".nomedia",
+        "license",
+        "contribute.md",
+        "readme.md",
+        "credits.md"
+    )
+        .map { Regex(it, RegexOption.IGNORE_CASE) }
+
+    private fun isValidModFolder(dir: FileHandle): Boolean {
+        var good = 0
+        var bad = 0
+        for (file in dir.list()) {
+            val goodList = if (file.isDirectory) goodFolders else goodFiles
+            if (goodList.any { file.name().matches(it) }) good++ else bad++
+        }
+        return good > 0 && good > bad
+    }
+
+    /** Check whether the unpacked zip contains a subfolder with mod content or is already the mod.
+     *  If there's a subfolder we'll assume it is the mod name, optionally suffixed with branch or release-tag name like github does.
+     *  @return Pair: actual mod content folder to name (subfolder name or [defaultModName])
+     */
+    private fun resolveZipStructure(
+        dir: FileHandle,
+        defaultModName: String
+    ): Pair<FileHandle, String> {
+        if (isValidModFolder(dir))
+            return dir to defaultModName
+        val subdirs =
+            dir.list(FileFilter { it.isDirectory })  // See detekt/#6822 - a direct lambda-to-SAM with typed `it` fails detektAnalysis
+        if (subdirs.size != 1 || !isValidModFolder(subdirs[0]))
+            throw UncivShowableException("Invalid Mod archive structure")
+        return subdirs[0] to choosePrettierName(subdirs[0].name(), defaultModName)
+    }
+
+    private fun choosePrettierName(folderName: String, defaultModName: String): String {
+        fun String.isMixedCase() = this != lowercase() && this != uppercase()
+        if (defaultModName.startsWith(
+                folderName,
+                true
+            ) && defaultModName.isMixedCase() && !folderName.isMixedCase()
+        )
+            return defaultModName.removeSuffix("-main").removeSuffix("-master")
+        return folderName
+    }
 }
