+++ core/src/com/unciv/ui/screens/savescreens/LoadOrSaveScreen.kt
@@ -13,6 +13,7 @@ import com.unciv.UncivGame
 import com.unciv.logic.UncivShowableException
 import com.unciv.logic.github.Github
 import com.unciv.logic.github.Github.folderNameToRepoName
+import com.unciv.logic.github.GithubAPI.downloadAndExtract
 import com.unciv.models.translations.tr
 import com.unciv.ui.components.UncivTooltip.Companion.addTooltip
 import com.unciv.ui.components.extensions.UncivDateFormat.formatDate
@@ -220,11 +221,9 @@ abstract class LoadOrSaveScreen(
                     ?: throw UncivShowableException("Could not download mod list.")
                 val repo = repos.items.firstOrNull { it.name.lowercase() == modName }
                     ?: throw UncivShowableException("Could not find a mod named \"[$modName]\".")
-                val modFolder = Github.downloadAndExtract(
-                    repo,
-                    UncivGame.Current.files.getModsFolder()
-                )
-                    ?: throw Exception("Unexpected 404 error") // downloadAndExtract returns null for 404 errors and the like -> display something!
+                val modFolder =
+                    repo.downloadAndExtract(UncivGame.Current.files.getModsFolder())
+                        ?: throw Exception("Unexpected 404 error") // downloadAndExtract returns null for 404 errors and the like -> display something!
                 Github.rewriteModOptions(repo, modFolder)
                 onModDownloaded(repo.name)
             }
